my-website/backend/vla_service/requirements.txt
# Core dependencies for VLA service
torch>=1.12.0
transformers>=4.20.0
openai-whisper
sounddevice
numpy>=1.21.0
scipy

# ROS 2 dependencies (for development)
rclpy
cv-bridge

# For Isaac Sim (if using the bridge)
omni-isaac-gym-envs
omni-isaac-orbit

# Additional utilities
pyyaml
requests